requirements.txt
Flask==2.0.2
azure-ai-formrecognizer==5.1.0
azure-core==1.19.0
gunicorn==21.2.0